+++ hscontrol/suite_test.go
@@ -1,7 +1,6 @@
 package hscontrol
 
 import (
-	"net/netip"
 	"os"
 	"testing"
 
@@ -46,9 +45,6 @@ func (s *Suite) ResetDB(c *check.C) {
 			Sqlite: types.SqliteConfig{
 				Path: tmpDir + "/headscale_test.db",
 			},
-		},
-		IPPrefixes: []netip.Prefix{
-			netip.MustParsePrefix("10.27.0.0/23"),
 		},
 		OIDC: types.OIDCConfig{
 			StripEmaildomain: false,